TSS/Eventlib: ifdef out non implemeted functions in skiboot
Skiboot doesn't implement fread, fwrite and sscanf and thus we need to
ifdef out functions or part of functions that usem these.

For TSS_EVENT2_Line_Read and TSS_EVENT_Line_Read both aren't used by
skiboot so it's easy to rule them out as dependent on the TPM_TSS_NOFILE
flag, which is set when compiling skiboot.

For TSS_SetCommandPort, TSS_SetPlatformPort, TSS_SetEncryptSessions and
TSS_Array_Scan TSS_Array_Scan it's also harmless to rule out the pieces of
code using sscanf as those are specific functions used only when sockets
and files are enabled.

Finally sscanf use in TSS_SetTraceLevel is ifdef out when TPM_SKIBOOT is
set as it was done for __ULTRAVISOR__ in
commit c8bac25 ("TSS: ifdef out sscanf()").

Signed-off-by: Mauro S. M. Rodrigues <maurosr@linux.vnet.ibm.com>