TSS/Eventlog: Adding multiple helpers for eventlib
This introduces multiple functions for help and deal with Evenlog data,
specially for  structures that belong to event data but also to other
structures, and thus it's adding similar code to deal with the little
endian used when marshaling/unmarshaling eventlog data to/from memory
buffer.

For instance like pcr extension operations doesn't use an specfic
endianess, but also marshall/unmarshal Algorithm ID, which is also present
at TSS_Event2 which needs to be LE when marshalling and unmarshalling

Also added helpers to calculate digest and event size.

Signed-off-by: Mauro S. M. Rodrigues <maurosr@linux.vnet.ibm.com>